// frontend/src/components/chatWindow/MessageContainer/messageContainer.module.scss.d.ts
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

declare const styles: {
  'message-main-container': string;
  'message-content-container': string;
  'input-container': string;
  'chat-input': string;
  'send-button': string;
  'input-container-wrapper': string;
  'container-padding-bottom': string;
  'advisoryTextContent': string;
  'inputDisabledTooltip': string;
  'popoverContainer': string;
};
export default styles;